refactor: WaitingRepository 경로 수정 및 import 정리
