+++ src/main/java/roomescape/reservation/domain/repository/WaitingRepository.java
@@ -1,8 +1,10 @@
-package roomescape.reservation.domain;
+package roomescape.reservation.domain.repository;
 
 import java.util.List;
 import java.util.Optional;
 import roomescape.member.domain.Member;
+import roomescape.reservation.domain.ReservationSlot;
+import roomescape.reservation.domain.Waiting;
 import roomescape.reservation.infrastructure.projection.WaitingWithRankProjection;
 
 public interface WaitingRepository {
