+++ src/test/java/roomescape/reservation/application/WaitingServiceTest.java
@@ -29,9 +29,9 @@ import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationSlot;
 import roomescape.reservation.domain.ReservationTime;
 import roomescape.reservation.domain.Waiting;
-import roomescape.reservation.domain.WaitingRepository;
 import roomescape.reservation.domain.repository.ReservationRepository;
 import roomescape.reservation.domain.repository.ReservationTimeRepository;
+import roomescape.reservation.domain.repository.WaitingRepository;
 import roomescape.reservation.ui.dto.request.CreateWaitingRequest;
 import roomescape.reservation.ui.dto.response.WaitingResponse;
 import roomescape.reservation.ui.dto.response.WaitingWithRankResponse;
